Refactor App component and introduce Layout component for improved structure and theme handling
